chore(ci): Experiment with auto-update (#117)
This PR trying to experiment with automated action to create a PR to upgrade swc_* when a new version is released. It is an action that runs periodically (2 times a day) which runs `cargo upgrade`, then create a PR if there's an update to be applied. If there is an open PR with a specific branch name, it won't create a new one but try to update the PR with a new commit when there's an update.

To create a fully working pull request, action requires personal access token, unfortunately. I choose the name `GHA_UPGRADE_TOKEN` for those having one under repository secrets will enable the action to perform:


We can use the default `GITHUB_TOKEN` still, but it'll create PR with limited access to github-actions bot and will not trigger any action to verify build is passing or not. (ref: https://github.com/peter-evans/create-pull-request/issues/48)

My goal is to let these actions run across transitive dependencies we own, so most of the upgrade process is simple enough to click the merge button unless there are breaking changes.