fix(emotion): Ensure that dirname is used over path (#155)
The current documentation and original emotion babel plugin documentation specify using `[dirname]` as a label token, however, the current implementation uses `[dir]`. I also found a case when testing where the parent would be selected with the remainder of the path if a full path was provided.

Resolve both of these, and add additional fixtures to test the output.