feat(transform-imports): Support mapping rules for individual import names (#196)
## Current Problem

Currently, we only have one `transform: '...'` rule applied to all
import names of a module import declaration. This implies that all
exports of that module will follow the same rule, which is usually not
the case.

Here're examples of some popular libraries:
-
[lucide-react](https://unpkg.com/browse/lucide-react@0.252.0/dist/esm/lucide-react.js):
`Accessibility`, `AccessibilityIcon` and `LucideAccessibility` all map
to `'/icons/accessibility'`
-
[react-bootstrap](https://unpkg.com/browse/react-bootstrap@2.8.0/esm/index.js):
special case for `useAccordionButton`
-
[@mui/material](https://unpkg.com/browse/@mui/material@5.14.1/index.js):
`export *` expressions

To solve these problems, the original Babel plugin [supports using a
function as the
transformer](https://www.npmjs.com/package/babel-plugin-transform-imports#using-a-function-as-the-transformer)
which brings the flexibility. We can't do the same but this PR adds an
(almost) equivalent API.

## API

Instead of using just a string for `transform`, this PR extends it to
accept an array of tuples that can match individual import names:

```js
'my-library/?(((\\w*)?/?)*)': {
  transform: [
    ["use(*)", "my-library/hooks/{{ kebabCase memberMatches.[1] }}"],
    ["(*)Icon", "my-library/icons/{{ memberMatches.[1] }}"],
    ["*", "my-library/{{ matches.[1] }}/{{ member }}"],
  ],
},
```

It will attempt to match each import in order and do the corresponding
transform. The new `memberMatches` value is similar to the existing
`matches` but it's from the imported name, rather than the module path.
All Handlebars helpers are still available.

With the configuration above, for example `import { useFooBar } from
'my-library/baz'` will be transpiled as `import { useFooBar } from
'my-library/hooks/foo-bar'`.

It's recommended to have a `*` condition by the end to match all
leftover cases. It can be pointing to the original lib path for
compatibility.

This new API is added in a non-breaking way so all existing code should
continue to work.

---------

Co-authored-by: Donny/강동윤 <kdy1997.dev@gmail.com>