fix(transform-imports): Transform side-effect imports (#473)
Closes: https://github.com/swc-project/plugins/issues/474

---

[Original issue](https://github.com/swc-project/plugins/issues/329):
Imports with side effects were being incorrectly removed. The expected
behavior is to retain the import and apply the transformation.

The current behavior causes issues when the code includes TypeScript
side-effect imports. These are skipped by the plugin and remain
untransformed, which leads to runtime failures in Node.js (since it
cannot import .ts files directly).