fix(formatjs): Fix ast tansform in jsx (#496)
- Ensure messages in function calls and JSX are transformed to AST
format when the ast option is set.
- Update tests and snapshots to reflect the correct AST output.

Fixes #497

Fixes an issue where messages were not being transformed to AST format
when the ast option was enabled.
Now, both function calls and JSX components correctly output the
defaultMessage as an AST when requested.

Also added a snapshot test for both `defineMessage`, `formatMessage` and
`FormattedMessage`.