fix(contrib/graphql): Add import below directives (#517)
Directives like "use strict", "use client" or "use cache" should always
be on top. Inserting an import statement on the first line breaks this.

For example with nextjs:
```
Ecmascript file had an error
> 1 | "use cache";
    | ^^^^^^^^^^^
  2 | import { cacheLife } from "next/dist/server/use-cache/cache-life";
  3 | import { cacheTag } from "next/dist/server/use-cache/cache-tag";
  4 | import { notFound } from "next/navigation";

The "use cache" directive must be at the top of the file.
```

This change will first scan if there is a line with "use ", and insert
the import below it.